# TestTorneo/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Torneos",
        "description": "Documentación de la API para gestionar torneos, jugadores y partidas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/jugadores": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Listar todos los jugadores",
                "description": "Permite listar todos los jugadores, opcionalmente filtrando por género.",
                "operationId": "dea173c276f606c29f946190cd3fbf0d",
                "parameters": [
                    {
                        "name": "genero",
                        "in": "query",
                        "description": "Filtrar por género (Masculino, Femenino, Todos)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de jugadores"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            },
            "post": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Crear un nuevo jugador",
                "description": "Registra un nuevo jugador con sus atributos.",
                "operationId": "a6bca232b4957d71a1179acf4eba7097",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "dni",
                                    "genero",
                                    "habilidad"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "genero": {
                                        "type": "string",
                                        "enum": [
                                            "Masculino",
                                            "Femenino"
                                        ]
                                    },
                                    "habilidad": {
                                        "type": "integer",
                                        "example": 85
                                    },
                                    "fuerza": {
                                        "type": "integer",
                                        "example": 80,
                                        "nullable": true
                                    },
                                    "velocidad": {
                                        "type": "integer",
                                        "example": 90,
                                        "nullable": true
                                    },
                                    "reaccion": {
                                        "type": "integer",
                                        "example": 70,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jugador creado exitosamente"
                    },
                    "400": {
                        "description": "Error en la validación"
                    },
                    "500": {
                        "description": "Error en el servidor"
                    }
                }
            }
        },
        "/api/jugadores/{id}": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Obtener un jugador por ID",
                "description": "Busca un jugador en base a su ID único.",
                "operationId": "38a1e42cd09832fbd502015f2a56d7d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador encontrado"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Actualizar un jugador",
                "description": "Permite modificar la información de un jugador ya existente.",
                "operationId": "43b9b9ac5454068b7a7befe507c8e46f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "habilidad": {
                                        "type": "integer",
                                        "example": 85
                                    },
                                    "fuerza": {
                                        "type": "integer",
                                        "example": 80,
                                        "nullable": true
                                    },
                                    "velocidad": {
                                        "type": "integer",
                                        "example": 90,
                                        "nullable": true
                                    },
                                    "reaccion": {
                                        "type": "integer",
                                        "example": 70,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jugador actualizado correctamente"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Eliminar un jugador",
                "description": "Realiza la eliminación lógica de un jugador por ID.",
                "operationId": "ad3b63cc99eed2fea1490aca50426404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador eliminado correctamente"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
        "/api/jugadores/{dni}": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Obtener un jugador por DNI",
                "description": "Busca un jugador por su DNI único.",
                "operationId": "fe88f5b3598c55962614f71d92a40dd4",
                "parameters": [
                    {
                        "name": "dni",
                        "in": "path",
                        "description": "DNI del jugador",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jugador encontrado"
                    },
                    "404": {
                        "description": "Jugador no encontrado"
                    }
                }
            }
        },
        "/api/jugadores/{id}/torneos": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Obtener torneos en los que participa un jugador",
                "description": "Obtiene la lista de torneos en los que ha participado un jugador. Se puede filtrar por torneos ganados.",
                "operationId": "86a5d25e083d1d67e5972819fee65a3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ganados",
                        "in": "query",
                        "description": "Filtrar solo torneos ganados (true, false, todos)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de torneos del jugador"
                    },
                    "404": {
                        "description": "Jugador no encontrado o sin torneos"
                    }
                }
            }
        },
        "/api/jugadores/{id}/partidas": {
            "get": {
                "tags": [
                    "Jugadores"
                ],
                "summary": "Obtener partidas en las que participa un jugador",
                "description": "Obtiene la lista de partidas en las que ha jugado un jugador, filtrando por ganadas, perdidas o todas.",
                "operationId": "9ab67c6248df5653c361a61540ac38f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del jugador",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "filtro",
                        "in": "query",
                        "description": "Filtrar por tipo de partida (ganadas, perdidas, todas)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de partidas del jugador"
                    },
                    "404": {
                        "description": "Jugador no encontrado o sin partidas"
                    }
                }
            }
        },
...
    },
...
}